First module of simulation class: 'heat' + test at the end.
Timestamp: 21.06.2018
This contribution is a first proposal to the simulation part of the udkm1Dsimpy. 

So far the class is initialized and takes THREE input arguments heat(structure,boundaryConditions,forceRecalc)
This is different to the MATLAB version, where only `structure` and `forceRecalc` were required. 
The change is due to the fact that now the boundary conditions are defined as an individual class, which gets passed on and not
by a structure which is created within the heat class. 
This can easily be changed!

Since not all of the functions in the structure class were ready at this point (21.06.2018) Some of the properties, which are provided
by these functions, are either `artificially` created, in order to proceede or the content conected to those has been left out.
If one wants to continue working on this class, one has to check for the latest update on the structure file too. 

Functions that have been taken into consideration: 
checkInitialTemperature
checkExcitation
getAbsorbtionProfile
getTemperatureAfterDeltaExcitation
generateSource
pdeBoundaryConditions

Functions that still need to be implemented: 
getHash
Display
getTempMap
calcTempMap 
calcHeatDiffusion
pdeHeatEquation