Class for boundary conditions
Timestamp: 21.06.2018
In order to implement the boundary conditions which are later needed in the `heat` class, especially in the funcition `calcHeatDiffusion`, an own class has been created.
In the MATLAB version this has been avoided by creating a strcture with several fields within the heat class. 
However, if doing it with a structure within the `heat class` at the end turns out to be the better option, one can easily change it to this way. 
Note: Now the `heat` class takes in addition to `structure` and `forceRecalc` one additional input argument, being the `boundaryConditions`.